+++ app/Http/Controllers/FreelancerController.php
@@ -123,8 +123,15 @@ class FreelancerController extends Controller
         // Verifica se o usuário tem permissão para deletar (usertype 2 ou 3)
         if ($user->usertype == 2 || $user->usertype == 3) {
             try {
-                Freelancer::destroy($id);
-                return redirect(route('dashboard'))->with('success', 'Registro deletado com sucesso');
+                $freelancer = Freelancer::findOrFail($id);
+                if ($freelancer->return_status != 'Em análise') {
+
+                    return redirect(route('dashboard'))
+                        ->with('fail', 'Consultas completas não podem ser excluídas.');
+                }
+                    Freelancer::destroy($id);
+                    return redirect(route('dashboard'))->with('success', 'Registro deletado com sucesso');
+                
             } catch (ValidationException $e) {
                 // Armazena os dados na sessão para depuração
                 return redirect(route('dashboard'))
